change: moved custom_title_filter_factory to core/utils/admin
